requirements.txt
bcrypt==3.2.0
blinker==1.4
Bootstrap-Flask==2.0.2
cffi==1.15.0
click==8.1.2
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.0
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Limiter==2.4.5.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-SeaSurf==1.1.1
Flask-SQLAlchemy==2.5.1
flask-talisman==1.0.0
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
is-safe-url==1.0
itsdangerous==2.1.2
Jinja2==3.1.1
limits==2.6.1
MarkupSafe==2.1.1
packaging==21.3
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pyparsing==3.0.8
rich==12.3.0
six==1.16.0
SQLAlchemy==1.4.36
typing_extensions==4.2.0
validate-email-address==1
Werkzeug==2.1.1
wrapt==1.14.0
WTForms==3.0.1
